Add https://sci-hub.bz/ to global test list.
sci-hub.bz is one of the Sci-Hub's remaining unblocked domains. The
other two that are already present in global.csv are sci-hub.org and
sci-hub.ac; the former was blocked in November 2015 and the latter was
ordered blocked on 2017-11-03 and actually blocked on 2017-11-17.
https://www.sciencemag.org/news/2017/11/court-demands-search-engines-and-internet-service-providers-block-sci-hub
https://www.theregister.co.uk/2017/11/23/scihubs_become_inactive_following_court_order/

I set the date_added for the entry as today's date. There previously was
an http version of the URL with date_added=2016-07-13. It was added in
commit 29d5b026689c766a43315c561fd33482077cfe52 and removed in commit
5da1b78c3e8e6389b78ad6d59f493aa6ca53acb6.

Other domains still unblocked (not added to the list):
https://sci-hub.la/
http://scihub22266oqcxt.onion/