Use a different strategy for archiving URLs
It seems like https://archive.is/submit/ when called from this script
will only archive the page if NO archive exists.

I am not sure if that is desireable (as we probably want a history of
the page as well).
Even passing in the {anyway: 1} parameter yields no effect.

Experimenting on the client I see that an extra submitid is passed.
Maybe that is the key, but I would rather not try to trick the good
people of the internet archive as it seems they don't want scripts
to archive things using that service.

On the other hand, web.archive.org/save/ seems to work just fine, so I
am switching over to using that for the time being.