Correct ill-formed dates in all lists.
In several of the lists, there are "date_added" values that aren't
parseable as ISO 8601 dates (that is, not in the form YYYY-MM-DD).
The most serious such case is a date recorded as "2018-14" with no
month, and there were also an entire file's worth of entries in the
ambiguous XX/XX/XXXX format.

This patch corrects all of the problem dates and also adds a check
for date validity to lint-lists.py.  I made it super picky; it will
even complain about "2017-3-1" instead of "2017-03-01".  This is to
make life as easy as possible for people trying to process the CSV
files in the future.  Whenever it was unclear what a date should
have been, I used the Git commit date as a guideline.