chore(it.csv): zap parked/nonexisting domains
This diff deletes a bunch of domains that we found to be parked
or nonexisting a while ago when testing web connectivity.

All these domains are part of the Italian test list.

They all have some residual censorship, but we've been favouring
existing websites recently, so I think this diff makes sense.

Closes https://github.com/ooni/probe/issues/2316

Closes https://github.com/ooni/probe/issues/2315

Closes https://github.com/ooni/probe/issues/2314

Closes https://github.com/ooni/probe/issues/2312

Closes https://github.com/ooni/probe/issues/2302

Closes https://github.com/ooni/probe/issues/2308

Closes https://github.com/ooni/probe/issues/2304

Closes https://github.com/ooni/probe/issues/2310