cleanup: use gardener dnsfix
This diff has been generated with gardener dnsfix. The gardener
is a new tool introduced in https://github.com/ooni/probe-cli/pull/1114.

The dnsfix command removes the domains that return errors using the
resolver we use in the TH _and_ for which we did not see anomalies
or confirmed cases in the ~last month.