Add use-application-dns.net - explicit signal for DNS filtering
This domain is meant to be used as a signal that overt DNS filtering is
in effect. Applications can try to resolve the name as a test for the
presence of DNS filtering. DNS filters that wish to announce their
presence can respond with NXDOMAIN rather than an IP address, to cause
some applications to disable DNS over HTTP support.

https://support.mozilla.org/en-US/kb/canary-domain-use-application-dnsnet
https://support.mozilla.org/en-US/kb/configuring-networks-disable-dns-over-https